Change capacity type to long

+++ src/main/java/de/uni_koblenz/west/koral/master/statisticsDB/impl/multi_file/storage/LRUCache.java
@@ -11,8 +11,8 @@ import java.util.Map.Entry;
 import java.util.Set;
 
 /**
- * A generic LRU cache with O(1) operations. Uses a doubly-linked-list for
- * access order plus an index (Map) for O(1) access.
+ * A generic LRU cache with O(1) operations. Uses a doubly-linked-list for access order plus an index (Map) for O(1)
+ * access.
  *
  * @author philipp
  *
@@ -21,7 +21,7 @@ import java.util.Set;
  */
 public class LRUCache<K, V> implements Iterable<Entry<K, V>> {
 
-	private final int capacity;
+	private final long capacity;
 
 	final Map<K, DoublyLinkedNode> index;
 
@@ -29,9 +29,9 @@ public class LRUCache<K, V> implements Iterable<Entry<K, V>> {
 
 	DoublyLinkedNode tail;
 
-	int size;
+	long size;
 
-	public LRUCache(int capacity) {
+	public LRUCache(long capacity) {
 		this.capacity = capacity;
 
 		index = new HashMap<>();
@@ -138,8 +138,7 @@ public class LRUCache<K, V> implements Iterable<Entry<K, V>> {
 	public Collection<Entry<K, V>> entrySet() {
 		LinkedList<Entry<K, V>> entrySet = new LinkedList<>();
 		for (Entry<K, DoublyLinkedNode> entry : index.entrySet()) {
-			entrySet.add(
-					new AbstractMap.SimpleImmutableEntry<>(entry.getKey(), entry.getValue().value));
+			entrySet.add(new AbstractMap.SimpleImmutableEntry<>(entry.getKey(), entry.getValue().value));
 		}
 		return entrySet;
 	}
@@ -157,8 +156,7 @@ public class LRUCache<K, V> implements Iterable<Entry<K, V>> {
 			@Override
 			public Entry<K, V> next() {
 				Entry<K, LRUCache<K, V>.DoublyLinkedNode> next = iterator.next();
-				Entry<K, V> result = new AbstractMap.SimpleImmutableEntry<>(next.getKey(),
-						next.getValue().value);
+				Entry<K, V> result = new AbstractMap.SimpleImmutableEntry<>(next.getKey(), next.getValue().value);
 				return result;
 			}
 		};
@@ -175,11 +173,10 @@ public class LRUCache<K, V> implements Iterable<Entry<K, V>> {
 	}
 
 	/**
-	 * @return The amount of elements in the cache, that is the doubly linked list.
-	 *         There may be more in the index map, depending on sub-implementations
-	 *         of {@link #removeEldest(DoublyLinkedNode)}.
+	 * @return The amount of elements in the cache, that is the doubly linked list. There may be more in the index map,
+	 *         depending on sub-implementations of {@link #removeEldest(DoublyLinkedNode)}.
 	 */
-	public int size() {
+	public long size() {
 		return size;
 	}
